fix: 941 - respect rate-limits during integration tests
New files:
* `too_many_requests_exception.dart`: Exception when the server returns "Too many requests".
* `too_many_requests_manager.dart`: Manager dedicated to "too many requests" server response.

Deleted files:
* `api_get_to_be_completed_products_test.dart`: moved "searchProduct" code to `api_search_products_test`
* `api_get_user_products_test.dart`: moved "searchProduct" code to `api_search_products_test`
* `api_matched_product_v2_test.dart`: moved "searchProduct" code to `api_search_products_test`

Impacted files:
* `api_get_localized_product_test.dart`: now using the `TooManyRequestsManager` for "getProduct" queries
* `api_get_product_image_ids_test.dart`: now using the `TooManyRequestsManager` for "getProduct" queries
* `api_get_product_test.dart`: moved "searchProduct" code to `api_search_products_test`; now using the `TooManyRequestsManager` for "getProduct" queries
* `api_json_to_from_test.dart`: now using the `TooManyRequestsManager` for "getProduct" queries
* `api_matched_product_v1_test.dart`: now using the `TooManyRequestsManager` for "getProduct" queries
* `api_not_food_get_product_test.dart`: now using the `TooManyRequestsManager` for "getProduct" queries
* `api_ocr_ingredients_test.dart`: now using the `TooManyRequestsManager` for "getProduct" queries
* `api_search_products_test.dart`: now using the `TooManyRequestsManager` for "searchProducts" queries; moved code there from other test files, gathering all "searchProducts" queries here
* `get_locations_order.dart`: unrelated minor improvement
* `get_proofs_order.dart`: unrelated minor improvement
* `open_food_api_client.dart`: now checking for `TooManyRequestsException`s for "getProduct" and "searchProducts" queries
* `openfoodfacts.dart`: added the 2 new files
