+++ lib/src/prices/get_locations_order.dart
@@ -2,6 +2,7 @@ import 'order_by.dart';
 
 /// Field for the "order by" clause of "get locations".
 enum GetLocationsOrderField implements OrderByField {
+  priceCount(offTag: 'price_count'),
   created(offTag: 'created'),
   updated(offTag: 'updated');
