+++ test/api_not_food_get_product_test.dart
@@ -34,6 +34,7 @@ void main() {
         fields: [ProductField.BARCODE],
         version: ProductQueryVersion(2),
       );
+      await getProductTooManyRequestsManager.waitIfNeeded();
       final OldProductResult result = await OpenFoodAPIClient.getOldProduct(
         configurations,
         uriHelper: uriHelper,
