+++ test/api_ocr_ingredients_test.dart
@@ -125,6 +125,7 @@ void main() {
         fields: [ProductField.INGREDIENTS_TEXT],
         version: ProductQueryVersion.v3,
       );
+      await getProductTooManyRequestsManager.waitIfNeeded();
       final ProductResultV3 result = await OpenFoodAPIClient.getProductV3(
         configurations,
         user: TestConstants.PROD_USER,
